# Conflicts:
#	test/api/example.cpp

+++ src/resources.hpp
@@ -6,9 +6,9 @@ namespace CaDiCaL {
 double absolute_real_time ();
 double absolute_process_time ();
 
-size_t maximum_resident_set_size ();
-size_t current_resident_set_size ();
+uint64_t maximum_resident_set_size ();
+uint64_t current_resident_set_size ();
 
-}
+} // namespace CaDiCaL
 
 #endif // ifndef _resources_hpp_INCLUDED
